epa_form/requirements.txt
flask==3.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
gspread==5.12.4
openai==1.12.0
python-dotenv==1.0.1 